+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The types of changes are:
 
 ### Added
 * Add delete confirmation when deleting a field or collection from a dataset [808](https://github.com/ethyca/fides/issues/808)
+* Add ability to delete datasets from the UI
 * Initial configuration wizard UI view
   * System scanning step: AWS credentials form and initial `generate` API usage.
 * Added Cypress for testing [713](https://github.com/ethyca/fides/pull/833)